move function to nav
